Combining 3 ContentExtensions files into 1
